backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
anthropic>=0.39
python-dotenv>=1.0
django-filter>=24.0
django-allauth>=65.0
requests>=2.28
PyJWT>=2.0
cryptography>=41.0
